added limit on RAM consumption to avoid slurm killing job due to out-of-memory
