+++ Converter/include/logger.h
@@ -9,8 +9,6 @@ using std::string;
 
 namespace logger {
 
-    extern int totalGridErrors;
-
 	void addOutputFile(string path);
 
 	void info(string msg, string file, int line);
@@ -18,8 +16,6 @@ namespace logger {
 	void warn(string msg, string file, int line);
 
 	void error(string msg, string file, int line);
-
-    std::ofstream& getLogFile();
 
 #define INFO(message) info(message, __FILE__, __LINE__)
 #define ERROR(message) error(message, __FILE__, __LINE__)
