+++ PotreeConverterBatchedConfigDelftblue.ini
@@ -139,7 +139,7 @@ LogFile = /home/anauman/escience/projects/nD-PC/PotreeConverter/logs/PotreeConve
 
 [INPUT_OUTPUT]
 InputDir = /home/anauman/staff-umbrella/ahn3
-OutputDir = /home/anauman/staff-umbrella/ahn3_PotreeConverterMPI_full_ahn3_full_mpi
+OutputDir = /home/anauman/staff-umbrella/PotreeConverterMPI_full_ahn3_full_mpi
 LazHeadersDir =/scratch/anauman/escience/projects/nD-PC/ahn3_headers
 PartitionsCSV = /home/anauman/escience/projects/nD-PC/PotreeConverter/ahn3_partitions_8x8.csv
 
@@ -150,9 +150,10 @@ CountingBatchSize = 20*(1024)**3
 DistributionBatchSize = 20*(1024)**3
 TmpDir = /scratch/anauman/escience/projects/nD-PC/PotreeConverterMPI_full_ahn3_full_mpi_tmp
 
+
 [SLURM_SCHEDULER]
 ProgramPath = /home/anauman/escience/projects/nD-PC/PotreeConverter/build/PotreeConverterMPI
 ProgramName = PotreeConverterMPI
-ConverterOptions = --encoding BROTLI --bounds [-60000.0,340000.0],[256250.0,656250.0],[-50000.0,350000.0]
+ConverterOptions = --encoding BROTLI --max-mem 130 --bounds [-60000.0,340000.0],[256250.0,656250.0],[-50000.0,350000.0]
 #Sbatch parameters are used to submit the job to the cluster. Note "%%" in --output and --error as a way to escape the "%" character in Python.
-SbatchParameters = --job-name=PotreeConverterMPI_full_ahn3_full_mpi --output=/home/anauman/escience/projects/nD-PC/PotreeConverter/sbatchOutput/PotreeConverterMPI_full_ahn3_full_mpi_%%j.out --error=/home/anauman/escience/projects/nD-PC/PotreeConverter/sbatchOutput/PotreeConverterMPI_full_ahn3_full_mpi_%%j.err --partition=compute --account=research-abe-aet --nodes=20 --ntasks-per-node=1 --cpus-per-task=12 --time=100:00:00 --mem=100G
+SbatchParameters = --job-name=PotreeConverterMPI_full_ahn3_full_mpi --output=/home/anauman/escience/projects/nD-PC/PotreeConverter/sbatchOutput/PotreeConverterMPI_full_ahn3_full_mpi_%%j.out --error=/home/anauman/escience/projects/nD-PC/PotreeConverter/sbatchOutput/PotreeConverterMPI_full_ahn3_full_mpi_%%j.err --partition=compute --account=research-abe-aet --nodes=20 --ntasks-per-node=1 --cpus-per-task=12 --time=100:00:00 --mem=150G
